refactor(kitchenpos): MenuFactory 삭제 및 MenuGroup - Menu 일대다 연관관계 삭제
